revamped gui.py, documented the other two
gui.py was built again from the ground up, as it was complete and utter shite.
game.py was documented with comments as to make it easier to navigate and connect to gui.py, although i've not yet done that.
boards.py was also documented with comments, as to know what does what, and to make the connections to gui.py via game.py easier. Still have to do that.